requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
Django==2.2
django-cors-headers==2.5.2
django-extensions==2.1.6
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
Markdown==2.6.11
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.2.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.6.4
pytz==2018.5
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.25
wcwidth==0.1.7
Werkzeug==0.15.2
